cambios de datos de la base de datos

+++ database/migrations/2021_11_19_201925_create_rentals_table.php
@@ -18,7 +18,7 @@ class CreateRentalsTable extends Migration
             $table->integer('active');
             $table->integer('days_rented');
             $table->integer('days_late')->nullable();
-            $table->integer('penalty_fee')->nullable();
+            $table->double('penalty_fee')->nullable();
             $table->timestamp('delivery_date')->nullable();
             $table->unsignedBigInteger('user_id');
             $table->unsignedBigInteger('movie_id');
